All interactable assets
Here's a lot of art! it should be all devices and items, as well as all customers and customer paws.

In addition there's an on-fire-doughnut sprite that I don't think we specfically need, but I had to make it to make the "on fire in the oil" sprite, so there's an extra sprite